+++ lib/Data/Time/Calendar/Gregorian.hs
@@ -179,9 +179,7 @@ instance Show Day where
     show = showGregorian
 
 -- orphan instance
-
--- | Year
-instance HasDays Year where
-    firstDayOf y = YearMonthDay y January 1
-    lastDayOf y = YearMonthDay y December 31
+instance DayPeriod Year where
+    periodFirstDay y = YearMonthDay y January 1
+    periodLastDay y = YearMonthDay y December 31
     dayPeriod (YearMonthDay y _ _) = y